refactor: 영속성 클래스 infrastructure 하위 패키지 이동
