+++ src/main/java/io/hhplus/tdd/point/infrastructure/database/PointHistory.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.infrastructure.database;
 
 public record PointHistory(
         long id,
