+++ src/main/java/io/hhplus/tdd/point/infrastructure/database/PointHistoryTable.java
@@ -1,13 +1,10 @@
-package io.hhplus.tdd.database;
+package io.hhplus.tdd.point.infrastructure.database;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import org.springframework.stereotype.Component;
-
-import io.hhplus.tdd.point.PointHistory;
-import io.hhplus.tdd.point.TransactionType;
 
 /**
  * 해당 Table 클래스는 변경하지 않고 공개된 API 만을 사용해 데이터를 제어합니다.
