+++ src/main/java/io/hhplus/tdd/point/infrastructure/database/UserPoint.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.infrastructure.database;
 
 public record UserPoint(
         long id,
