+++ src/main/java/io/hhplus/tdd/point/infrastructure/database/UserPointRepository.java
@@ -1,6 +1,4 @@
-package io.hhplus.tdd.database;
-
-import io.hhplus.tdd.point.UserPoint;
+package io.hhplus.tdd.point.infrastructure.database;
 
 public interface UserPointRepository {
 	UserPoint selectById(Long id);
